requirements.txt
streamlit
Python-dotenv
PyPDF2
langchain-community
transformers
faiss-cpu
sentence-transformers
accelerate
langchain-huggingface
bitsandbytes